requirements.txt
Flask==1.1.2
waitress==1.4.4
requests==2.24.0
xlrd==1.2.0
loguru==0.5.3
python-dateutil==2.8.1
flask_sqlalchemy==2.5.1
pymysql==0.10.1
openpyxl==3.0.7
WTForms==2.3.3
Flask-WTF==0.14.3